requirements.txt
langchain-community>=0.0.10
tavily-python>=0.3.0
python-dotenv>=1.0.0 